Moved dispatcher to a separate file and added a comment.
